# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PlayOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      summary: Sign up as a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login as a player (email + password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== HOST AUTH ====================
...
  /bookings/my:
    get:
      operationId: getMyBookings
      tags: [bookings]
      summary: Get current player's bookings with game details
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyBookingsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelBookingResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not your booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SignUpBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        name:
          type: string
      required:
        - email
        - password
        - phone
        - name
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
...
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        name:
          type: string
        role:
          type: string
          enum: [player, organiser, host, admin]
        token:
          type: string
          description: JWT auth token returned on login/signup only
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - createdAt

    # ==================== GAME SCHEMAS ====================
...
    MyBookingGame:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        pitchName:
          type: string
        kickoffTime:
          type: string
          format: date-time
        price:
          type: number
        capacity:
          type: integer
        status:
          type: string
          enum: [open, full, cancelled]
        bookedCount:
          type: integer
      required:
        - id
        - title
        - pitchName
        - kickoffTime
        - price
        - capacity
        - status
        - bookedCount
    MyBooking:
      type: object
      properties:
        id:
          type: string
        gameId:
          type: string
        team:
          type: integer
        slotIndex:
          type: integer
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        bookedAt:
          type: string
          format: date-time
        game:
          $ref: "#/components/schemas/MyBookingGame"
      required:
        - id
        - gameId
        - team
        - slotIndex
        - paymentStatus
        - bookedAt
        - game
    MyBookingsResponse:
      type: object
      properties:
        upcoming:
          type: array
          items:
            $ref: "#/components/schemas/MyBooking"
        past:
          type: array
          items:
            $ref: "#/components/schemas/MyBooking"
      required:
        - upcoming
        - past
    CancelBookingResponse:
      type: object
      properties:
        success:
          type: boolean
        refundTier:
          type: string
          enum: [full, half, none]
        refundAmount:
          type: number
        message:
          type: string
      required:
        - success
        - refundTier
        - refundAmount
        - message

    # ==================== BOOKING SCHEMAS ====================
...