# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PlayOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from "./authUserRole";

export interface AuthUser {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  name: string;
  role: AuthUserRole;
  /** JWT auth token returned on login/signup only */
  token?: string;
  createdAt: Date;
}